requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
